refactor(22.5): separate utility functions in their own folder
