+++ back-end/bloco-22-introducao-ao-desenvolvimento-web-com-nodejs/dia-5-express-middlewares/exercise-1/inputValidationMiddleware.js
@@ -1,16 +1,4 @@
-function isUsernameValid(username) {
-  return Boolean(username && username.length > 3);
-}
-
-function isPasswordValid(password) {
-  return Boolean(password && password.length >= 4 && password.length <= 8
-    && !isNaN(password));
-}
-
-function isEmailValid(email) {
-  const emailRegex = /\b[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}\b/i;
-  return emailRegex.test(email);
-}
+const { isUsernameValid, isPasswordValid, isEmailValid } = require('./util');
 
 function inputValidationMiddleware(req, res, next) {
   const { username, email, password } = req.body;
